# docs/swagger.yaml
basePath: /
definitions:
  blocks.Block:
    properties:
      created_at:
        type: string
      data:
        type: string
      difficulty:
        type: integer
      hash:
        type: string
      id:
        type: integer
      id_user:
        type: string
      last_validation_date:
        type: string
      mined_at:
        type: string
      mined_by:
        type: string
      nonce:
        type: integer
      prev_hash:
        type: string
      status_id:
        type: integer
      timestamp:
        type: string
      updated_at:
        type: string
    type: object
...
  blocks.ResAllBlocks:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/blocks.Block'
        type: array
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
...
  miner.DataBlockToMine:
    properties:
      cuota:
        type: number
      data:
        items:
          type: integer
        type: array
      difficulty:
        type: integer
      id:
        type: integer
      prev_hash:
        items:
          type: integer
        type: array
      timestamp:
        type: string
    type: object
  miner.responseGetBlock:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/miner.DataBlockToMine'
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
...
  miner.responseRegisterMined:
    properties:
      code:
        type: integer
      data:
        type: string
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  miner.rqRegisterMined:
    properties:
      difficulty:
        type: integer
      hash:
        type: string
      nonce:
        type: integer
      wallet_id:
        type: string
    type: object
...
host: http://172.174.77.149:2054
info:
  contact:
    email: info@bjungle.net
    name: API Support
  description: Documentación del API que conecta con el core de BLion
  license:
    name: Software Owner
    url: https://www.bjungle.net/terms/licenses
  termsOfService: https://www.bjungle.net/terms/
  title: BLion
  version: "1.1"
paths:
...
  /api/v1/block/get-all/{limit}/{offset}:
    get:
      consumes:
      - application/json
      description: Método para obtener todos los bloques de la Blockchain de BLion
      parameters:
      - description: Número de bloques por petición
        in: path
        name: limit
        required: true
        type: string
      - description: Salto de bloques por petición
        in: path
        name: offset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blocks.ResAllBlocks'
      summary: Método para obtener todos los bloques de la Blockchain
      tags:
      - Block
  /api/v1/miner/block-to-mine:
    get:
      consumes:
      - application/json
      description: Método para obtener el bloque a minar
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/miner.responseGetBlock'
      summary: Método para obtener el bloque a minar
      tags:
      - Miner
...
  /api/v1/miner/register-mined:
    post:
      consumes:
      - application/json
      description: Método para registrar el hash del bloque minado
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos del hash del bloque minado
        in: body
        name: rqRegisterMined
        required: true
        schema:
          $ref: '#/definitions/miner.rqRegisterMined'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/miner.responseRegisterMined'
      summary: Método para registrar el hash del bloque minado
      tags:
      - Miner
...
swagger: "2.0"
...